requirements.txt
asgiref==3.2.3
boto3==1.10.39
botocore==1.13.39
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.40.0
django-cors-headers==3.2.0
django-haystack==2.8.1
django-storages==1.8
docutils==0.15.2
idna==2.8
jmespath==0.9.4
oauthlib==3.1.0
Pillow==6.2.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
Whoosh==2.7.4
